+++ NEWS.md
@@ -1,3 +1,10 @@
+# bcputility 0.1.1
+
+* rowterminator and fieldterminator are no passed to `data.table::fwrite` if data
+is in memory
+
+* Added support for geometry/geography data import for 'sf' objects
+
 # bcputility 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.